Display Groups baed on Location and interests / Refactoring getCategories
Groups are loaded on the Groups Page in a Ion Card layout after the user submits location and / or interested categories.

The getCatgories method has been removed in Settings and replaced by a loader on page load. The user gets a visual confirmation that the categories are loading, whilst waiting.