Refactoring Groups Page
Form has been removed, making way for a far more simple with use of a button to search for groups.

Image at the top of the ion-card has been added, showing an image uploaded by the group, if it's available.

If there's no categories, the Delete all Caregories Button will no longer display